+++ settings/LoanPolicySettings.js
@@ -1,13 +1,6 @@
 import React, { PropTypes } from 'react';
-import _ from 'lodash';
 
-import Paneset from '@folio/stripes-components/lib/Paneset';
-import Pane from '@folio/stripes-components/lib/Pane';
-import PaneMenu from '@folio/stripes-components/lib/PaneMenu';
-import Icon from '@folio/stripes-components/lib/Icon';
-import NavList from '@folio/stripes-components/lib/NavList';
-import NavListSection from '@folio/stripes-components/lib/NavListSection';
-
+import EntrySelector from './EntrySelector';
 import LoanPolicyDetail from './LoanPolicyDetail';
 
 class LoanPolicySettings extends React.Component {
@@ -34,42 +27,7 @@ class LoanPolicySettings extends React.Component {
   constructor(props) {
     super(props);
 
-    this.state = {
-      selectedPolicy: null,
-    };
-
-    this.clearSelection = this.clearSelection.bind(this);
-    this.onSelectRow = this.onSelectRow.bind(this);
     this.createNewPolicy = this.createNewPolicy.bind(this);
-  }
-
-  componentDidUpdate(prevProps) {
-    // Follows example from permission set settings in ui-users
-    const policyDiffs = _.differenceBy(this.props.data.loanPolicies, prevProps.data.loanPolicies, 'id');
-    const newPolicy = policyDiffs[0];
-    //console.log("misc: pc", newPolicy.pendingCreate)
-
-    if (newPolicy) {
-      // At this point in the lifecycle the CID is still on the object, and
-      // this messes up the saveing of the Permission Set. It should not be needed any longer
-      // and will be removed.
-      delete newPolicy._cid; // eslint-disable-line no-underscore-dangle
-
-      // eslint-disable-next-line react/no-did-update-set-state
-      this.setState({
-        selectedPolicy: newPolicy,
-      });
-    }
-  }
-
-  onSelectRow(id, e) {
-    e.preventDefault();
-    this.setState({ selectedPolicy: _.find(this.props.data.loanPolicies, { id }) });
-  }
-
-  clearSelection() {
-    const policyForFocus = this.props.data.loanPolicies ? _.sortBy(this.props.data.loanPolicies, ['name'])[0] : '';
-    this.setState({ selectedPolicy: policyForFocus });
   }
 
   createNewPolicy() {
@@ -88,40 +46,41 @@ class LoanPolicySettings extends React.Component {
       },
     });
   }
+  
+  componentDidUpdate(prevProps) {
+    // Follows example from permission set settings in ui-users
+    const policyDiffs = _.differenceBy(this.props.data.loanPolicies, prevProps.data.loanPolicies, 'id');
+    const newPolicy = policyDiffs[0];
+    //console.log("misc: pc", newPolicy.pendingCreate)
+
+    if (newPolicy) {
+      // At this point in the lifecycle the CID is still on the object, and
+      // this messes up the saveing of the Permission Set. It should not be needed any longer
+      // and will be removed.
+      delete newPolicy._cid; // eslint-disable-line no-underscore-dangle
+
+      // eslint-disable-next-line react/no-did-update-set-state
+      // this.setState({
+      //   selectedPolicy: newPolicy,
+      // });
+      this.props.history.push(`${this.props.match.path}/${newPolicy.id}`)
+    }
+  }
 
   render() {
-    const { data } = this.props;
-    const policies = _.sortBy(data.loanPolicies, ['name']);
-    const policyDisplay = policies != null ? policies.map(p =>
-      <a
-        key={p.id}
-        href={`#${p.id}`}
-        onClick={this.onSelectRow.bind(this, p.id)}
-      >
-        {p.name ? p.name : 'Untitled loan policy'}
-      </a>) : null;
-
-    const LoanPolicyLastMenu = (
-      <PaneMenu>
-        <button title="Add loan policy" onClick={this.createNewPolicy}>
-          <Icon icon="plus-sign" />
-        </button>
-      </PaneMenu>
-    );
+    const policies = _.sortBy(this.props.data.loanPolicies, ['name']);
 
     return (
-      <Paneset nested>
-        <Pane defaultWidth="25%" lastMenu={LoanPolicyLastMenu} paneTitle={this.props.label}>
-          <NavList>
-            <NavListSection activeLink={this.state.selectedPolicy ? `#${this.state.selectedPolicy.id}` : ''}>
-              {policyDisplay}
-            </NavListSection>
-          </NavList>
-        </Pane>
-        {this.state.selectedPolicy && <Pane paneTitle={this.state.selectedPolicy.name} defaultWidth="fill">
-          <LoanPolicyDetail initialValues={this.state.selectedPolicy} parentMutator={this.props.mutator} clearSelection={this.clearSelection} />
-        </Pane>}
-      </Paneset>
+    
+        <EntrySelector
+          {...this.props}
+          detailComponent={LoanPolicyDetail}
+          allEntries={policies}
+          entryCreator={this.createNewPolicy}
+          parentMutator={this.props.mutator}
+          paneTitle="Loan policies"
+          addButtonTitle="Add loan policy"
+        />
     );
   }